refactoring service and controller methods

+++ Complains/src/main/java/com/fiap/hackathon/complains/service/impl/ComplainsServiceImpl.java
@@ -3,25 +3,29 @@ package com.fiap.hackathon.complains.service.impl;
 import java.util.*;
 import java.util.stream.Collectors;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.fiap.hackathon.complains.exception.ResourceNotFoundException;
+import com.fiap.hackathon.complains.helper.ComplainsHelper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBSaveExpression;
-import com.amazonaws.services.dynamodbv2.model.AttributeValue;
-import com.amazonaws.services.dynamodbv2.model.ComparisonOperator;
-import com.amazonaws.services.dynamodbv2.model.ExpectedAttributeValue;
-import com.fiap.hackathon.complains.dto.ComplainsDTO;
-import com.fiap.hackathon.complains.dto.NovaComplainDTO;
-import com.fiap.hackathon.complains.entity.Complains;
+import com.fiap.hackathon.complains.model.dto.ComplainsDTO;
+import com.fiap.hackathon.complains.model.dto.NovaComplainDTO;
+import com.fiap.hackathon.complains.model.entity.Complains;
 import com.fiap.hackathon.complains.repository.ComplainsRepository;
 import com.fiap.hackathon.complains.service.ComplainsService;
+import org.springframework.web.client.HttpClientErrorException;
+
+import static com.fiap.hackathon.complains.helper.ComplainsHelper.complainsDTOBuilder;
+import static com.fiap.hackathon.complains.helper.ComplainsHelper.complainsUpdateBuilder;
+import static com.fiap.hackathon.complains.helper.ComplainsHelper.createComplainsBuilder;
 
 @Service
 public class ComplainsServiceImpl implements ComplainsService {
 	
 	private ComplainsRepository complainsRepository;
-	
+
+	@Autowired
 	public ComplainsServiceImpl(ComplainsRepository complainsRepository) {
 		this.complainsRepository = complainsRepository;
 	}
@@ -32,36 +36,38 @@ public class ComplainsServiceImpl implements ComplainsService {
 		complainsList = (List<Complains>) complainsRepository.findAll();
 		return complainsList
 				.stream()
-				.map(ComplainsDTO::new)
+				.map(ComplainsHelper::complainsDTOBuilder)
 				.collect(Collectors.toList());
 	}
 
 	@Override
 	public ComplainsDTO buscarComplainPorId(String id) {
-		Optional<Complains> complains = complainsRepository.findById(id);
-		return new ComplainsDTO(complains);
+		Complains complains = complainsRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(HttpStatus.NOT_FOUND, "Id Not found!"));
+		return complainsDTOBuilder(complains);
 	}
 
 	@Override
 	public ComplainsDTO criar(NovaComplainDTO novaComplainDTO) {
-		Complains complains = new Complains(novaComplainDTO);
-		Complains savedComplain = complainsRepository.save(new Complains(null, new Date(),new Date(),"usuário joao"));
-		return new ComplainsDTO(savedComplain);
+		Complains savedComplain = complainsRepository.save(createComplainsBuilder(novaComplainDTO));
+		return complainsDTOBuilder(savedComplain);
 	}
 
 	@Override
 	public ComplainsDTO atualizar(String id, NovaComplainDTO novaComplainDTO) {
-		Optional<Complains> complains = complainsRepository.findById(id); 
-		complains.orElseThrow().setId(novaComplainDTO.getId());
-		complains.orElseThrow().setDataAlteracao(GregorianCalendar.getInstance().getTime());
-		complains.orElseThrow().setUsuario(novaComplainDTO.getUsuario());
-		
-		Complains updatedComplain = complainsRepository.save(complains.get());
-		return new ComplainsDTO(updatedComplain);
+		Complains complains = complainsRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(HttpStatus.NOT_FOUND, "Id Not found!"));
+		Complains complainsUpdated = complainsUpdateBuilder(complains, novaComplainDTO);
+		Complains updatedComplain = complainsRepository.save(complainsUpdated);
+		return complainsDTOBuilder(updatedComplain);
 	}
 
 	@Override
 	public void deletarComplain(String id) {
+		Complains complains = complainsRepository.findById(id).orElseThrow(() -> new HttpClientErrorException(HttpStatus.NOT_FOUND, "Id Not found!"));
+
+		if(Objects.isNull(complains)){
+			throw new HttpClientErrorException(HttpStatus.NOT_FOUND, "Id Not found!");
+		}
+
 		complainsRepository.deleteById(id);
 	}
